refactor: move implementation of Program.UpmPack() -> Upm.Context.PackUpmPackage()
#copypasta
